requirements.txt
asgiref==3.5.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
ConfigArgParse==1.5.3
Django==4.0.3
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
Faker==13.3.3
Flask==2.0.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
geventhttpclient==1.5.3
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.0
locust==2.8.4
MarkupSafe==2.1.1
msgpack==1.0.3
psutil==5.9.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.1
pywin32==303
pyzmq==22.3.0
requests==2.27.1
roundrobin==0.0.2
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
Werkzeug==2.0.3
zope.event==4.5.0
zope.interface==5.4.0
